# Conflicts:
#	.gitignore
#	android/app/src/crypto/java/updated/mysterium/vpn/ui/top/up/amount/usd/TopUpAmountUsdViewModel.kt
#	android/app/src/crypto/java/updated/mysterium/vpn/ui/top/up/crypto/payment/CryptoPaymentActivity.kt
#	android/app/src/crypto/res/drawable/icon_stripe_payment_option.png
#	android/app/src/main/java/updated/mysterium/vpn/ui/top/up/card/summary/CardSummaryActivity.kt
#	android/app/src/main/java/updated/mysterium/vpn/ui/top/up/coingate/amount/TopUpAmountActivity.kt
#	android/app/src/main/res/values-night/colors.xml
#	android/app/src/main/res/values/colors.xml
#	android/app/src/main/res/values/strings.xml

+++ android/app/src/main/java/updated/mysterium/vpn/model/payment/OrderOptions.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class OrderOptions(
     val minimum: Double,
     @SerializedName("suggested")
-    val amountsSuggestion: List<Int>
+    val amountsSuggestion: List<Double>
 )
